[CEDS-2978]_Implement_verified_email_check
Add action to most pages to check that the
current user has a verified email address
associated with their EORI number.

If it is detected that the user does not
have a verified email address then they
are redirected to a page notifying them
of the fact and providing them with a link
to the customs-email-frontend service to
rectify the situation.